Bringing mapping improvements from Partners back to Involvements.
Closes #62
Closes #63
Closes #44
